+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 Yii2 Queue Extension Change Log
 ===============================
 
-2.3.7 under development
------------------------
-- Enh #509: Add StatisticsProviderInterface to get statistics from queue (kalmer)
+2.3.7 April 29, 2024
+--------------------
+
 - no changes in this release.
+- Enh #509: Add StatisticsProviderInterface to get statistics from queue (kalmer)
 
 
 2.3.6 October 03, 2023